+++ tests/app/main/views/test_add_service.py
@@ -4,6 +4,16 @@ import app
 from app.utils import is_gov_user
 
 
+def test_non_gov_user_cannot_see_add_service_button(client,
+                                                    mock_login,
+                                                    mock_get_non_govuser,
+                                                    api_nongov_user_active):
+    client.login(api_nongov_user_active)
+    response = client.get(url_for('main.choose_service'))
+    assert 'Add a new service' not in response.get_data(as_text=True)
+    assert response.status_code == 200
+
+
 def test_get_should_render_add_service_template(app_,
                                                 api_user_active,
                                                 mocker):
